moved useEffect to Posts and fixed duplicate loading problem
when going back to all posts after searching, some posts were duplicated
